Jakub's movement fixes. Changed floats to doubles in collision/pathfinding calculations

+++ extension/movement.h
@@ -4,6 +4,6 @@
 typedef struct GameState GameState;
 typedef struct Entity Entity;
 
-void move(GameState* state, Entity** entity, float dt);
+void move(GameState* state, Entity** entity, double dt);
 
 #endif // MOVEMENT_H
